fix imports; move `file_utils` to `indexer.utils`
